+++ src/games/mafia/Announcment.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
-import UserApi from './UserApi.js'
+import UserApi from '../../UserApi.js'
 
 export default class Announcment extends Component {
 
     render() {
-        
+
         console.log(this.props.killedPlayers);
-        
+
         var listItems = this.props.killedPlayers.map((id) => {
             return (<li>{UserApi.getName(id)}</li>);
         });
