+++ src/games/mafia/Mafia.js
